a2a_multi_agent/pyproject.toml
[project]
name = "version_3_multi_agent"
version = "0.1.0"
description = "version_3_multi_agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncclick>=8.1.8",
    "click>=8.1.8",
    "fastapi>=0.115.12",
    "openai-agents>=0.1.0",
    "openai>=1.0.0",
    "httpx>=0.28.1",
    "httpx-sse>=0.4.0",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "starlette>=0.46.2",
    "uvicorn>=0.34.2",
]
